requirements.txt
numpy==1.19.5
matplotlib==3.3.4
fastapi==0.68.1
Pillow==8.3.2
uvicorn[standard]
